+++ talker_flutter/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.5.1
+- Update display message UI
+- Update talker version to 0.6.1
+
 ## 0.5.0
 - Implement UI filter for convenient work with the console in the application